requirements.txt
flet 
pandas 
pywin32 
pillow


